moved components to separate folder and added css files
