+++ front-sample/src/components/AllUserInfo.css
@@ -1,0 +1,8 @@
+.AllUserInfo {
+    max-width: 35%;
+    height: 300px;
+    float: left;
+    padding: 10px;
+  
+    background-color: lightgreen;  
+}
