+++ front-sample/src/components/AllUserInfo.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactTable from 'react-table';
 import 'react-table/react-table.css';
+import './AllUserInfo.css';
 
 class AllUserInfo extends React.Component {
   constructor() {
