+++ front-sample/src/components/ProductList.css
@@ -1,0 +1,8 @@
+.ProductList {
+    max-width: 30%;
+    height: 300px;
+    float: left;
+    padding: 10px;
+  
+    background-color: lightblue;
+}
