+++ front-sample/src/components/ProductList.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactTable from 'react-table';
-import 'react-table/react-table.css'
+import 'react-table/react-table.css';
+import './ProductList.css';
 
 class ProductList extends React.Component {
   constructor() {
