+++ front-sample/src/components/UserInfo.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import './UserInfo.css';
 
 class UserInfo extends React.Component {
   constructor() {
